Add ΔE2000 and CAM16UCS metrics. Code only right now. Easy experiments with `contrast -.35 EY` show a bunch of background-14 entries all easier to read than the infamous bg0,fg4 dark blue on black.  The very same is true of contrast -.35 CY.
This new -THRESH mode does suggest a way to maybe score all these
scoring methods with manageable effort, though:
  - A) Get social agreement 0,4 is too hard to read
  - B) Find the score of that via each scoring method
  - C) Use the program to dump the tables
  - D) Count the "exceptions", namely the number of cells subjectively
       easier to read than 0,4.
That would systematically count all the false negatives (scored as
harder to read than they should be), and a table could be made of all
the various methods.

A similar procedure can count false positives (scored as easier to read
than they should), but that has a lot more numbers in the tables/visual
noise and so is more human work.